requirements.txt
loguru==0.6.0
python-dotenv==1.0.0
requests==2.31.0
yadisk==1.3.3
python-telegram-bot==13.7
tqdm~=4.66.1